esp32/requirements.txt
micropython-esp32-i2c
micropython-esp32-rtc
umqtt.simple
